info al user sobre instalación del plugin

+++ register.cs
@@ -4,6 +4,7 @@ using System.Reflection;
 using Autodesk.AutoCAD.Runtime;
 using Autodesk.AutoCAD.ApplicationServices;
 using Autodesk.AutoCAD.DatabaseServices;
+using System.Windows.Forms;
 
 namespace pruebaAcadForm
 {
@@ -28,6 +29,7 @@ namespace pruebaAcadForm
                 if (subKey.Equals(sAppName))
                 {
                     regAcadAppKey.Close();
+                    MessageBox.Show("Ya está registrado el plugin");
                     return;
                 }
             }
@@ -41,7 +43,7 @@ namespace pruebaAcadForm
             regAppAddInKey.SetValue("LOADCTRLS", 14, RegistryValueKind.DWord);
             regAppAddInKey.SetValue("LOADER", sAssemblyPath, RegistryValueKind.String);
             regAppAddInKey.SetValue("MANAGED", 1, RegistryValueKind.DWord);
-
+            MessageBox.Show("Plugin registrado OK!");
             regAcadAppKey.Close();
         }
 
@@ -55,9 +57,27 @@ namespace pruebaAcadForm
             Autodesk.AutoCAD.Runtime.RegistryKey regAcadProdKey = Autodesk.AutoCAD.Runtime.Registry.CurrentUser.OpenSubKey(sProdKey);
             Autodesk.AutoCAD.Runtime.RegistryKey regAcadAppKey = regAcadProdKey.OpenSubKey("Applications", true);
 
-            // Delete the key for the application
-            regAcadAppKey.DeleteSubKeyTree(sAppName);
-            regAcadAppKey.Close();
+            // Check to see if the "MyApp" key exists
+            bool appUnRegistered = false;
+            string[] subKeys = regAcadAppKey.GetSubKeyNames();
+            foreach (string subKey in subKeys)
+            {
+                // If the application is already registered, unregister
+                if (subKey.Equals(sAppName))
+                {
+                    // Delete the key for the application
+                    regAcadAppKey.DeleteSubKeyTree(sAppName);
+                    regAcadAppKey.Close();
+                    MessageBox.Show("Plugin desregistrado!");
+                    appUnRegistered = true;
+                    return;
+                }
+            }
+            if (!appUnRegistered) 
+            {
+                MessageBox.Show("El Plugin no está registrado!");
+            }
+
         }
     }
 }
